backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
pymongo[srv]==4.5.0
